refactor : 중복 이름 예외 로직 위치 변경
