+++ src/main/java/blackjack/domain/BlackJackGame.java
@@ -2,9 +2,12 @@ package blackjack.domain;
 
 import static blackjack.domain.Player.*;
 
+import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 public class BlackJackGame {
@@ -17,15 +20,31 @@ public class BlackJackGame {
 		this.gamers = gamers;
 	}
 
-	public void setInitialCards(final Deck deck) {
-		giveTwoCards(dealer, deck);
-
+	public static BlackJackGame initializeSetting(final List<String> names, final Deck deck) {
+		Player dealer = new Dealer();
+		List<Gamer> gamers = createGamers(names);
+		setInitialCards(dealer, deck);
 		for (Gamer gamer : gamers) {
-			giveTwoCards(gamer, deck);
+			setInitialCards(gamer, deck);
+		}
+		return new BlackJackGame(dealer, gamers);
+	}
+
+	private static List<Gamer> createGamers(List<String> names) {
+		checkDuplicateName(names);
+		return names.stream()
+			.map(Gamer::new)
+			.collect(Collectors.toList());
+	}
+
+	private static void checkDuplicateName(final List<String> names) {
+		Set<String> removalDuplicateNames = new HashSet<>(names);
+		if (names.size() != removalDuplicateNames.size()) {
+			throw new IllegalArgumentException("[ERROR] 중복된 이름은 입력할 수 없습니다.");
 		}
 	}
 
-	private void giveTwoCards(final Player player, final Deck deck) {
+	private static void setInitialCards(final Player player, final Deck deck) {
 		for (int i = 0; i < PLAYER_ALREADY_CARD_SIZE; i++) {
 			player.receiveCard(deck.draw());
 		}
@@ -35,10 +54,8 @@ public class BlackJackGame {
 		int dealerResult = dealer.calculateResult();
 		return gamers.stream()
 			.collect(Collectors.toMap(gamer -> gamer,
-				gamer -> Result.findResult(dealerResult,
-					gamer.calculateResult()),
-				(e1, e2) -> e1,
-				LinkedHashMap::new));
+				gamer -> Result.findResult(dealerResult, gamer.calculateResult()),
+				(e1, e2) -> e1, LinkedHashMap::new));
 	}
 
 	public Player getDealer() {
